+++ crates/client/db/src/lib.rs
@@ -146,8 +146,6 @@ pub enum Column {
     /// Meta column for block storage (sync tip, pending block)
     BlockStorageMeta,
 
-    RevertedTxs,
-
     /// Contract class hash to class data
     ContractClassData,
 
@@ -210,7 +208,6 @@ impl Column {
             TxHashToBlockN,
             BlockHashToBlockN,
             BlockStorageMeta,
-            RevertedTxs,
             ContractClassData,
             ContractToClassHashes,
             ContractToNonces,
@@ -239,7 +236,6 @@ impl Column {
             TxHashToBlockN => "tx_hash_to_block_n",
             BlockHashToBlockN => "block_hash_to_block_n",
             BlockStorageMeta => "block_storage_meta",
-            RevertedTxs => "reverted_txs",
             BonsaiContractsTrie => "bonsai_contracts_trie",
             BonsaiContractsFlat => "bonsai_contracts_flat",
             BonsaiContractsLog => "bonsai_contracts_log",