+++ crates/client/sync/src/commitments/mod.rs
@@ -7,12 +7,13 @@ use blockifier::state::cached_state::CommitmentStateDiff;
 use classes::class_trie_root;
 use contracts::contract_trie_root;
 use dc_db::DeoxysBackend;
-use dp_convert::to_stark_felt::ToStarkFelt;
+use dp_convert::ToStarkFelt;
+use dp_receipt::Event;
 use events::memory_event_commitment;
 use indexmap::IndexMap;
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
 use starknet_api::hash::StarkFelt;
-use starknet_api::transaction::{Event, Transaction};
+use starknet_api::transaction::Transaction;
 use starknet_core::types::{
     ContractStorageDiffItem, DeclaredClassItem, DeployedContractItem, NonceUpdate, ReplacedClassItem, StateUpdate,
     StorageEntry,