+++ crates/primitives/block/src/header.rs
@@ -3,8 +3,8 @@ use core::num::NonZeroU128;
 use blockifier::block::{BlockInfo, GasPrices};
 use blockifier::context::{BlockContext, ChainInfo, FeeTokenAddresses};
 use blockifier::versioned_constants::VersionedConstants;
-use dp_convert::to_felt::ToFelt;
-use primitive_types::U256;
+use dp_convert::ToFelt;
+use dp_transactions::MAIN_CHAIN_ID;
 use starknet_api::block::{BlockNumber, BlockTimestamp};
 use starknet_api::core::{ChainId, ContractAddress};
 use starknet_api::data_availability::L1DataAvailabilityMode;
@@ -12,6 +12,8 @@ use starknet_api::hash::StarkHash;
 use starknet_types_core::felt::Felt;
 use starknet_types_core::hash::Pedersen;
 use starknet_types_core::hash::StarkHash as StarkHashTrait;
+
+use crate::starknet_version::StarknetVersion;
 
 /// Block status.
 ///
@@ -59,13 +61,11 @@ pub struct Header {
     /// A commitment to the events produced in this block
     pub event_commitment: StarkHash,
     /// The version of the Starknet protocol used when creating this block
-    pub protocol_version: String, // TODO: Verify if the type can be changed to u8 for the protocol version
+    pub protocol_version: StarknetVersion,
     /// Gas prices for this block
     pub l1_gas_price: Option<GasPrices>,
     /// The mode of data availability for this block
     pub l1_da_mode: L1DataAvailabilityMode,
-    /// Extraneous data that might be useful for running transactions
-    pub extra_data: Option<U256>,
 }
 
 const BLOCKIFIER_VERSIONED_CONSTANTS_JSON_0_13_0: &[u8] = include_bytes!("../resources/versioned_constants_13_0.json");
@@ -94,10 +94,9 @@ impl Header {
         transaction_commitment: StarkHash,
         event_count: u128,
         event_commitment: StarkHash,
-        protocol_version: String,
+        protocol_version: StarknetVersion,
         gas_prices: Option<GasPrices>,
         l1_da_mode: L1DataAvailabilityMode,
-        extra_data: Option<U256>,
     ) -> Self {
         Self {
             parent_block_hash,
@@ -112,15 +111,21 @@ impl Header {
             protocol_version,
             l1_gas_price: gas_prices,
             l1_da_mode,
-            extra_data,
         }
     }
 
     /// Converts to a blockifier BlockContext
-    pub fn into_block_context(&self, fee_token_addresses: FeeTokenAddresses, chain_id: ChainId) -> BlockContext {
-        let versioned_constants = if self.block_number <= 607_877 {
+    pub fn into_block_context(
+        &self,
+        fee_token_addresses: FeeTokenAddresses,
+        chain_id: ChainId,
+    ) -> Result<BlockContext, &'static str> {
+        let protocol_version = self.protocol_version;
+        let versioned_constants = if protocol_version < StarknetVersion::STARKNET_VERSION_0_13_0 {
+            return Err("Unsupported protocol version");
+        } else if protocol_version < StarknetVersion::STARKNET_VERSION_0_13_1 {
             &BLOCKIFIER_VERSIONED_CONSTANTS_0_13_0
-        } else if self.block_number <= 632_914 {
+        } else if protocol_version < StarknetVersion::STARKNET_VERSION_0_13_1_1 {
             &BLOCKIFIER_VERSIONED_CONSTANTS_0_13_1
         } else {
             VersionedConstants::latest_constants()
@@ -142,31 +147,13 @@ impl Header {
             use_kzg_da: self.l1_da_mode == L1DataAvailabilityMode::Blob,
         };
 
-        BlockContext::new_unchecked(&block_info, &chain_info, versioned_constants)
+        Ok(BlockContext::new_unchecked(&block_info, &chain_info, versioned_constants))
     }
 
     /// Compute the hash of the header.
-    pub fn hash(&self) -> Felt {
-        if self.block_number >= 833 {
-            // Computes the block hash for blocks generated after Cairo 0.7.0
-            let data: &[Felt] = &[
-                Felt::from(self.block_number),         // block number
-                self.global_state_root.to_felt(),      // global state root
-                self.sequencer_address.to_felt(),      // sequencer address
-                Felt::from(self.block_timestamp),      // block timestamp
-                Felt::from(self.transaction_count),    // number of transactions
-                self.transaction_commitment.to_felt(), // transaction commitment
-                Felt::from(self.event_count),          // number of events
-                self.event_commitment.to_felt(),       // event commitment
-                Felt::ZERO,                            // reserved: protocol version
-                Felt::ZERO,                            // reserved: extra data
-                self.parent_block_hash.to_felt(),      // parent block hash
-            ];
-
-            Pedersen::hash_array(data)
-        } else {
-            // Computes the block hash for blocks generated before Cairo 0.7.0
-            let data: &[Felt] = &[
+    pub fn hash(&self, chain_id: Felt) -> Felt {
+        if self.block_number < 833 && chain_id == MAIN_CHAIN_ID {
+            Pedersen::hash_array(&[
                 Felt::from(self.block_number),
                 self.global_state_root.to_felt(),
                 Felt::ZERO,
@@ -179,9 +166,22 @@ impl Header {
                 Felt::ZERO,
                 Felt::from_bytes_be_slice(b"SN_MAIN"),
                 self.parent_block_hash.to_felt(),
-            ];
-
-            Pedersen::hash_array(data)
+            ])
+        } else {
+            // Computes the block hash for blocks generated after Cairo 0.7.0
+            Pedersen::hash_array(&[
+                Felt::from(self.block_number),         // block number
+                self.global_state_root.to_felt(),      // global state root
+                self.sequencer_address.to_felt(),      // sequencer address
+                Felt::from(self.block_timestamp),      // block timestamp
+                Felt::from(self.transaction_count),    // number of transactions
+                self.transaction_commitment.to_felt(), // transaction commitment
+                Felt::from(self.event_count),          // number of events
+                self.event_commitment.to_felt(),       // event commitment
+                Felt::ZERO,                            // reserved: protocol version
+                Felt::ZERO,                            // reserved: extra data
+                self.parent_block_hash.to_felt(),      // parent block hash
+            ])
         }
     }
 }